+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+- Unarchive command to extract files into S3.
+
 ## 0.1.0
 
 - Initial release of the `cobalt-s3-archiver` crate.